[refactor] 마이페이지 닉네임 변경 리팩토링
